Added WorldEdit selection, support for different sized screens
WorldEdit selection can now be used for the screen by typing /jumbo set.

Also, view direction has been changed so that the word refers to the
direction player is facing when viewing the screen.

LED display size is now set in config and is distinct from in-game
screen size. Screen will be cropped to the top left if larger, or
centred in the display if smaller.